Agente/requirements.txt
fastapi
uvicorn
openai
llama-index
pandas
python-dotenv
requests
pytz